+++ test/e2e/pages/terminal.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { expect, Locator } from '@playwright/test';
-import { Code } from '../automation/code';
+import { Code } from '../infra/code';
 import { QuickAccess } from './quickaccess';
 
 const TERMINAL_WRAPPER = '#terminal .terminal-wrapper';
