+++ test/e2e/tests/console/console-history.test.ts
@@ -3,7 +3,7 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Application } from '../../automation';
+import { Application } from '../../infra';
 import { test, tags } from '../_test.setup';
 
 test.use({
