+++ test/e2e/tests/new-project-wizard/new-project-python.test.ts
@@ -3,7 +3,7 @@
  *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { PythonFixtures, ProjectType, ProjectWizardNavigateAction } from '../../automation';
+import { PythonFixtures, ProjectType, ProjectWizardNavigateAction } from '../../infra';
 import { test, expect, tags } from '../_test.setup';
 
 test.use({
